test: Move FP unit test helpers to a header file
