+++ test/unittests/CMakeLists.txt
@@ -20,6 +20,7 @@ target_sources(
     execute_control_test.cpp
     execute_death_test.cpp
     execute_floating_point_test.cpp
+    execute_floating_point_test.hpp
     execute_numeric_test.cpp
     execute_test.cpp
     floating_point_utils_test.cpp
